% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_utils.R
\name{add_bfdr}
\alias{add_bfdr}
\title{Add Bayesian FDR values to a limma top table}
\usage{
add_bfdr(ttab)
}
\arguments{
\item{ttab}{A data frame returned by \code{\link[limma:toptable]{limma::topTable()}} with
\code{n=Inf} containing B-statistics.}
}
\value{
The same table with additional columns named "PP" and
"BFDR".

Note that limma only produces B-statistics for single-contrast
tests (i.e. moderated t-tests). As a convenience, if this fuction
is called on a data frame with no column named "B", it will issue a
warning and return the data frame unmodified.
}
\description{
This function uses \code{\link[=bfdr]{bfdr()}} to compute the posterior probability of
differential expression and the FDR based on limma's B-statistics
and add them as columns named "PP" and "BFDR" respectively.
}
\examples{

# TODO copy examples from limma::topTable()

}
\seealso{
\code{\link[limma:toptable]{limma::topTable()}}.
}
