% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{add_numbered_colnames}
\alias{add_numbered_colnames}
\title{Add numbered colnames with a commo prefix.}
\usage{
add_numbered_colnames(x, prefix = "C")
}
\description{
Add numbered colnames with a commo prefix.
}
\examples{
# TODO: PCA example
}
