% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{arg_default}
\alias{arg_default}
\title{Emulate default arguments from the body of a function}
\usage{
arg_default(...)
}
\description{
This is useful if you need to do a requireNamespace before you can
evaluate the default.
}
