% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{auto_factorize_columns}
\alias{auto_factorize_columns}
\title{Intelligently convert character columns to factors in a data frame}
\usage{
auto_factorize_columns(df)
}
\description{
For each column of a data frame, if it is a character vector with
at least one repeated value, convert that column to a factor.
Character columns with all unique values (e.g. sample IDs) will not
be modified.
}
