% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{BPselectModel}
\alias{BPselectModel}
\title{Parallelized version of \code{\link[limma:selectModel]{limma::selectModel()}}}
\usage{
BPselectModel(y, designlist, criterion = "aic", df.prior = 0,
  s2.prior = NULL, s2.true = NULL, ..., BPPARAM = BiocParallel::bpparam())
}
\description{
Parallelized version of \code{\link[limma:selectModel]{limma::selectModel()}}
}
