% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{collapse_to_atomic}
\alias{collapse_to_atomic}
\title{Convert a list to an atomic vector.}
\usage{
collapse_to_atomic(x, sep = ",")
}
\arguments{
\item{x}{The list to convert.}

\item{sep}{Separator to use for collapsing vectors to scalars.}
}
\value{
An atomic vector that same length and names as \code{x}.
}
\description{
If \code{x} is already an atomc vector, it is not modified. If it is a
list, non-scalar elements are converted to scalars, and these are
concatenated to produce an atomic vector the same length as \code{x}.
Specifically, list elements with 0 length are replaced by NA, and
elements with length greater than one are pasted together into
strings using \code{sep} as the separator.
}
