% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{datetime_to_date}
\alias{datetime_to_date}
\title{Convert all-midnight POSIXt to Date}
\usage{
datetime_to_date(x, force = FALSE, eps_seconds = 1)
}
\arguments{
\item{x}{A POSIXt vector.}

\item{force}{If TRUE, then forcibly convert \code{x} to a Date even if its times
are not all midnight. This also speeds up the conversion by skipping the
check, so it mayu be useful if you already know that all times are
midnight.}

\item{eps_seconds}{"Fudge factor" for detecting midnight. As long as every
date-time is no more than this many seconds after midnight, it is
considered equivalent to midnight.}
}
\value{
Either a Date vector if the vector was converted, or the original
vector if not.
}
\description{
Dates in databases are often stored as date-time objects (POSIXt in R
parlance) with all the time values set to midnight (i.e. 00:00:00.0). This
function detects this case and converts them back to just Dates. Date-time
objects containing at least one non-midnight time are assumed to be genuine
date-times and are returned unmodified.
}
\examples{
library(lubridate)

actually_a_date <- ymd_hms("2000-01-01 00:00:00")
print(actually_a_date)
class(actually_a_date)
actually_a_date <- datetime_to_date(actually_a_date)
print(actually_a_date)
class(actually_a_date)

actually_a_datetime <- ymd_hms("2000-01-01 01:27:13")
print(actually_a_datetime)
class(actually_a_datetime)
actually_a_datetime <- datetime_to_date(actually_a_datetime)
print(actually_a_datetime)
class(actually_a_datetime)
}
