% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{datetimes_to_dates}
\alias{datetimes_to_dates}
\title{Convert all-midnight POSIXt columns to Date columns in a data frame}
\usage{
datetimes_to_dates(df, ...)
}
\arguments{
\item{df}{A data frame, potentially containing POSIXt columns.}

\item{...}{Additional arguments to \code{\link[=datetime_to_date]{datetime_to_date()}}. All other columns
are left alone.}
}
\value{
The original data frame, with all POSIXt columns mutated by
\code{\link[=datetime_to_date]{datetime_to_date()}}.
}
\description{
Convert all-midnight POSIXt columns to Date columns in a data frame
}
\examples{
library(lubridate)

mydf <- data.frame(
    chr_col = "a",
    num_col = 1,
    fac_col = factor(letters)[1],
    actually_a_date = ymd_hms("2000-01-01 00:00:00"),
    actually_a_datetime = ymd_hms("2000-01-01 01:27:13")
)
sapply(mydf, \(x) class(x)[1])
sapply(datetimes_to_dates(mydf), \(x) class(x)[1])
}
