% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{ensure_atomic_columns}
\alias{ensure_atomic_columns}
\title{Ensure that all columns of a data frame are atomic vectors.}
\usage{
ensure_atomic_columns(df, sep = ",")
}
\arguments{
\item{df}{The data frame whose list columns should be collapsed}

\item{sep}{The separator to use when collapsing vectors to scalars}
}
\value{
\code{df} with all columns converted to atomic vectors.
}
\description{
Columns that are lists have each of their elements collapsed into
strings using the sepcified separator.
}
\examples{

library(S4Vectors)
x <- DataFrame(a = 1:5, b = List(split(1:25, 1:5)))
ensure_atomic_columns(x)

}
\seealso{
\code{\link[=collapse_to_atomic]{collapse_to_atomic()}}
}
