% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{ensureAtomicColumns}
\alias{ensureAtomicColumns}
\title{Ensure that all columns of a data frame are atomic vectors.}
\usage{
ensureAtomicColumns(df, sep = ",")
}
\description{
Columns that are lists have each of their elements collapsed into
strings using the sepcified separator. TODO: If sep is NA, throw an
error for lists.
}
