% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{featureCountsParallel}
\alias{featureCountsParallel}
\title{Parallel version of \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}}}
\usage{
featureCountsParallel(files, ...)
}
\description{
Parallel version of \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}}
}
