% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports.R
\name{find_in_environment}
\alias{find_in_environment}
\title{Find the parent environment containing a name.}
\usage{
find_in_environment(
  what,
  envir,
  mode = "any",
  stop_after = .BaseNamespaceEnv,
  namespace_only = TRUE
)
}
\arguments{
\item{what}{A character string, the name to search for}

\item{envir}{The environment at which to start the search. If the name is not
found in this environment, then its parent (enclosing) environments are
searched. This can also be a function, in which case the function's
environment is used as the starting point.}

\item{mode}{a character string; if not \code{"any"}, only objects whose
    \code{\link{mode}} equals \code{mode} are searched.}

\item{stop_after}{An environment after which the search should be terminated.
The parent and further ancestors of this environment will not be searched.}

\item{namespace_only}{If TRUE, then only return the name of the found
environment if it is a namespace. Otherwise return NA.}
}
\value{
The name of the environment in which the specified name is found or
NA if the environment is unnamed, or the name is not found in any
environment.
}
\description{
Find the parent environment containing a name.
}
\examples{
find_in_environment("median", envir = stats::median)
find_in_environment("stop", envir = stats::median)
}
