% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{find_multi_value_groups}
\alias{find_multi_value_groups}
\title{Find groups with more than 1 row in them.}
\usage{
find_multi_value_groups(.data, ...)
}
\arguments{
\item{...}{Specification of column-names, the combination of which is
expected to be unique. Specified as in \code{\link[dplyr:group_by]{dplyr::group_by()}}.}
}
\value{
The subset of the original data frame, filtered to only rows in
groups containing 2 or more rows, sorted by the grouping columns.
}
\description{
This function is meant for finding unexpected exceptions to uniqueness
constraints on a set of columns.
}
\examples{
mydf <- data.frame(
    unique_id = rep(letters[1:3], c(1, 1, 3)),
    value = 1:5
)
# Uh-oh, this "unique" ID isn't unique
print(mydf)
# Find the ID values that aren't unique
find_multi_value_groups(mydf, unique_id)
}
