% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{format_bp}
\alias{format_bp}
\title{Format a number of base pairs using the most appropriate unit.}
\usage{
format_bp(x)
}
\arguments{
\item{x}{A numeric vector representing numbers of base pairs.}
}
\value{
A character vector containing string representations of the
number of base pairs, using appropriate SI prefixes.
}
\description{
Format a number of base pairs using the most appropriate unit.
}
\examples{

format_bp(c(100, 1000, 1e6))

}
