% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{future.lapply}
\alias{future.lapply}
\title{Like lapply, but returns a list of future objects.}
\usage{
future.lapply(X, FUN, ...)
}
\arguments{
\item{X, FUN, ...}{See \code{\link[=lapply]{lapply()}}.}
}
\value{
A list of futures. You can fetch all the results in a list
using \code{\link[future:values]{future::values()}}.
}
\description{
Like lapply, but returns a list of future objects.
}
\seealso{
\code{\link[=lapply]{lapply()}}, \code{\link[future:future]{future::future()}}
}
