% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports.R
\name{get_imports}
\alias{get_imports}
\title{Get a table of all imported functions in a function body}
\usage{
get_imports(
  fun,
  envir = environment(fun),
  ignore_env_names = c("base", "R_GlobalEnv", "R_EmptyEnv", NA)
)
}
\arguments{
\item{fun}{The function to search for imported functions}

\item{ignore_env_names}{List of package names to ignore. By default, "base"
is ignored since base functions do not need to be declared in packages. The
global and empty environments are also ignored, since functions in those
environments are not importable.}
}
\value{
A data frame listing all functions imported by \code{fun}, along with
various ways to import those functions into your package.
}
\description{
Get a table of all imported functions in a function body
}
\examples{

get_imports(stats::xtabs)

}
