% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_utils.R
\name{get_pval_colname}
\alias{get_pval_colname}
\title{Determine the column name of the p-value column in a table}
\usage{
get_pval_colname(ttab)
}
\arguments{
\item{ttab}{A table of results produced by \code{\link[limma:toptable]{limma::topTable()}},
\code{\link[edgeR:topTags]{edgeR::topTags()}}, \code{\link[DESeq2:results]{DESeq2::results()}}, or any table with a
p-value column in it.}
}
\value{
The name of the column containing p-values. If no such
column is found, an error is thrown.
}
\description{
limma uses "P.Value" while edgeR uses "PValue", so we need an
abstraction.
}
