% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_utils.R
\name{getBCVTable}
\alias{getBCVTable}
\title{Get a table of genewise BCV values from a DGEList}
\usage{
getBCVTable(y, design, ..., rawdisp)
}
\arguments{
\item{y}{A DGEList. If it does not already contain dispersion
values, they will be estimated.}

\item{design}{A design matrix to use when estimating dispersions.
This is optional if \code{y} already contains a design.}

\item{...}{Additional arguments are passed to
\code{\link[edgeR:estimateDisp]{edgeR::estimateDisp()}}, if dispersion estimation is required
(see below).}

\item{rawdisp}{Another DGEList, which should be identical to \code{y}
except that it had dispersions estimated with \code{prior.count} =
0. This argument is completely optional, and is only used for
speeding up the function if the raw dispersions have already
been estimated. To suppress estimation of raw dispersions
entirely, set this to NA, NULL, or FALSE.

(TODO: Take a long hard look at the logic in this function and make
sure it matches this documentation.)

If \code{y} already contains estimated dispersions, then no other
arguments are required, and the existing dispersion values will be
used. If \code{y} does not contain dispersions, then they will be
estimated using \code{design} and \code{...}. Similarly, \code{design} is only
required if \code{y} does not already contain a design (check if
\code{y$design} is non-NULL).}
}
\description{
Given a DGEList object, this will extract the average logCPM as
well as the common, trended, and genewise dispersion values. In
addition, it will re-estimate the dispersions with no empirical
Bayes shrinkage (prior.df = 0) to obtain the raw genewise
dispersion values.
}
