% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plots.R
\name{ggduo.dataXY}
\alias{ggduo.dataXY}
\title{Variant of \code{GGally::ggduo()} with separate arguments for  \code{dataX} and \code{dataY}}
\usage{
ggduo.dataXY(dataX, dataY, extraData = NULL, ...)
}
\description{
Variant of \code{GGally::ggduo()} with separate arguments for  \code{dataX} and \code{dataY}
}
