% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{is_valueless}
\alias{is_valueless}
\title{Determine whether a list or vector has any non-missing values.}
\usage{
is_valueless(x, empty_values = NA)
}
\arguments{
\item{x}{The object to check for emptiness.}

\item{empty_values}{A vector of values that should be considered
"empty", in addition to NA. Any vector containing only NA and
these values will be considered "valueless". For example, if
you want to know whether a numeric vector contains any non-NA,
non-zero values, you could use \code{empty_values = 0}. Another
common value to use is \code{""}, the empty string.}
}
\description{
A vector is "valueless" if it has zero length or contains only NA
values. A list is empty if has zero length or contains only empty
vectors.
}
\seealso{
\code{\link[rlang:is_empty]{rlang::is_empty()}}, which only checks for zero length.
}
