% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{liftOver_motifMap}
\alias{liftOver_motifMap}
\title{Convenience function for running liftOver on a MotifMap BED file}
\usage{
liftOver_motifMap(infile, chainfile, outfile, allow.gap = 2)
}
\description{
Convenience function for running liftOver on a MotifMap BED file
}
