% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{load_filtered}
\alias{load_filtered}
\title{Variant of \code{load()} that allows excluding specific names}
\usage{
load_filtered(file, envir = parent.frame(), ..., exclude = NULL)
}
\arguments{
\item{file, envir, ...}{These arguments have the same meaning as in
\code{load()}.}

\item{exclude}{Character vector of variable names not to load.}
}
\description{
If you want to load every object in an R data file except for a
specific set of variable names that you wish to exclude, this
function makes it easy to do so.
}
\examples{

x <- 5
y <- 3
datafile <- tempfile(fileext = ".rda")
# Only saves y, not x
save.image(datafile)
# Set new values which can be overridden by loading the old ones.
x <- 10
y <- 10
load_filtered(datafile, exclude = "x")
# x is still 10, but y is now 3 again
mget(c("x", "y"))

}
\seealso{
\code{\link[=load]{load()}}
}
