% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{mutate_if_present}
\alias{mutate_if_present}
\title{Perform mutations only if specific column names are present}
\usage{
mutate_if_present(.data, names, ...)
}
\arguments{
\item{.data, ...}{These have the same meaning as in
\code{\link[dplyr:mutate]{dplyr::mutate()}}}

\item{names}{If \code{.data} does not contain all of these names as
columns, it will be returned unmodified.}
}
\description{
Note that columns required to be present need not be involved in the mutations.
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}}. Not to be confused with
\code{\link[dplyr:mutate_if]{dplyr::mutate_if()}}, which mutates columns that match a
specific filter criteria.
}
