% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_ext_remove_all}
\alias{path_ext_remove_all}
\title{Remove all file extensions using \code{fs::path_ext_remove()} repeatedly.}
\usage{
path_ext_remove_all(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\value{
\code{path} with all extensions removed.
}
\description{
This function runs \code{\link[fs:path_ext_remove]{fs::path_ext_remove()}} as many times as needed to remove
all file extensions from \code{path}.
}
\examples{
fs::path_ext_remove("a.csv.gz")
path_ext_remove_all("a.csv.gz")
}
