% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plots.R
\name{plot_pval_hist}
\alias{plot_pval_hist}
\title{Create an annotatied p-value histogram plot}
\usage{
plot_pval_hist(pvals, nbins = 100, ptn = limma::propTrueNull)
}
\arguments{
\item{pvals}{A vector of p-values}

\item{nbins}{Number of bins in the histogram}

\item{ptn}{Estimated proportion of true null hypotheses. If not
provided, it will be estimated using \code{\link[limma:propTrueNull]{limma::propTrueNull()}}.
This can also be a function that takes one argument (the vector
of p-values) and returns the estimated proportion of true null
hypotheses, which should be a single number between 0 and 1.}
}
\value{
a ggplot object
}
\description{
This function takes a vector of p-values and plots a histogram of
them using ggplot. It will add horizontal lines at 1 and the
estimated proportion of true null hypotheses for reference.
}
