% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{print_var_vector}
\alias{print_var_vector}
\title{Print a readable summary of a list of values.}
\usage{
print_var_vector(v)
}
\description{
This is useful for printing out a list of the parsed command-line
arguments for a script.
}
\details{
TODO: Find a better home for this?
}
