% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_utils.R
\name{propTrueNullByLocalFDR_unrestricted}
\alias{propTrueNullByLocalFDR_unrestricted}
\title{Variant of propTrueNullByLocalFDR that does not restrict to the range \link{0,1}}
\usage{
propTrueNullByLocalFDR_unrestricted(p)
}
\description{
Useful for identifying potential atypical p-value distributions
(e.g. too many large p-values).
}
