% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{quotemeta}
\alias{quotemeta}
\title{Escape a string for use in a Perl-compatible regular expression}
\usage{
quotemeta(string)
}
\arguments{
\item{string}{The string to escape}
}
\value{
\code{string}, with all non-word characters escaped with backslashes.

Note that this strategy does not work for R's own regular
expression syntax, so you should specify \code{perl = TRUE}, or an
equivalent option, when using this to construct a regular
expression.
}
\description{
This is an implementation of Perl's \code{quotemeta} operator. It is
meant for escaping any characters that might have special meanings
in a PCRE.
}
\examples{

s <- "A (string) with [many] regex *characters* in it."
grepl(s, s, perl = TRUE)
quotemeta(s)
# After escaping, the string matches itself.
grepl(quotemeta(s), s, perl = TRUE)

}
