% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctutils-package.R, R/zzz.R
\docType{package}
\name{rctutils-package}
\alias{rctutils-package}
\title{rctutils: Utility functions by Ryan C. Thompson}
\description{
A collection of utility functions generally relating to bioinformatics analysis.

This is just here to tell roxygen2 about all base package imports,
which were recommended by R CMD check, as well as some common
imports that are used in many functions. Adding these to every
individual function that uses these common functions is too
tedious, so I've just added them all here.
}
\details{
TODO Convert these to import_from if possible. Also eliminate magrittr?
}
\author{
\strong{Maintainer}: Ryan C. Thompson \email{rct@thompsonclan.org}

Authors:
\itemize{
  \item Ryan C. Thompson \email{rct@thompsonclan.org}
}

}
\keyword{internal}
