% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{read.motifmap}
\alias{read.motifmap}
\title{Read MotifMap-provided BED file into a GRanges object.}
\usage{
read.motifmap(x, parse_name = TRUE)
}
\description{
We can't use rtracklayer::import.bed because it chokes on spaces in
fields, which MotifMap contains.
}
