% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{read.RDS.or.RDA}
\alias{read.RDS.or.RDA}
\title{Read a single object from RDS or RDA file}
\usage{
read.RDS.or.RDA(filename, expected.class = "ANY")
}
\description{
Read a single object from RDS or RDA file
}
