% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{read.single.object.from.rda}
\alias{read.single.object.from.rda}
\title{Read a single R object from an RDA file.}
\usage{
read.single.object.from.rda(filename)
}
\description{
If run on an RDA file containing more than one object, throws an
error.
}
