% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{read_narrowPeak}
\alias{read_narrowPeak}
\title{Read a narrowPeak format BED file.}
\usage{
read_narrowPeak(file, ...)
}
\arguments{
\item{file}{The file name to read from.}

\item{...}{Additional arguments are passed to \code{read.table()}.}
}
\value{
A GRanges object containing the ranges from \code{file}.
}
\description{
NarrowPeak is a sub-format of the BED file format used as the
output format of some peak callers. \code{\link[rtracklayer:BEDFile-class]{rtracklayer::import.bed()}}
often throws an error when reading these files, so a separate
function is needed.
}
