% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{read_regions}
\alias{read_regions}
\title{Read a GRanges from a variety of possible file types.}
\usage{
read_regions(filename)
}
\arguments{
\item{filename}{The file to read regions from.}
}
\value{
a GRanges object.

Possible input file types include: RDA or RDS file (containing a
GRanges or data frame), narrowPeak, BED, GFF, and SAF.
}
\description{
This function attempts to read in a set of genomic regions as a
GRanges object. It does its best to auto-detect the input file
type. If it cannot do so, it throws an error.
}
