% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{read_single_object_from_rda}
\alias{read_single_object_from_rda}
\title{Read a single R object from an RDA file.}
\usage{
read_single_object_from_rda(filename)
}
\description{
If run on an RDA file containing more than one object, throws an
error.
}
