% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{read_table_general}
\alias{read_table_general}
\title{General function for reading a table from any table-like file}
\usage{
read_table_general(filename, read.table.args = NULL,
  read.xlsx.args = NULL, dataframe.class = "data.frame")
}
\description{
General function for reading a table from any table-like file
}
