% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{read_tx2gene_from_genemap}
\alias{read_tx2gene_from_genemap}
\title{Read a Salmon "geneMap" file into a data frame}
\usage{
read_tx2gene_from_genemap(fname)
}
\arguments{
\item{fname}{The file to read from. This should generally be a file
named "genemap.txt" in a Salmon index directory}
}
\value{
A data frame with 2 columns named "TXNAME" and "GENEID".

Since this uses \code{\link[=read_table_general]{read_table_general()}} and allows additional
columns to be present beyond the first two, it is actually quite
lenient about the format of the table.
}
\description{
This function reads a file in the Salmon "geneMap" format into a
data frame. The format of this file is described in salmon's help
text for the "--geneMap" command-line option.
}
\seealso{
The "--geneMap" option to salmon, whose help text
describes the format of the file.
}
