% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{readsToFragmentMidpoints}
\alias{readsToFragmentMidpoints}
\title{Compute the midpoints of a collection of read fragments}
\usage{
readsToFragmentMidpoints(reads, fraglen,
  extend_fragment_upstream = FALSE)
}
\arguments{
\item{reads}{A \link[GenomicAlignments:GAlignments]{GenomicAlignments::GAlignments},
\link[GenomicAlignments:GAlignmentPairs]{GenomicAlignments::GAlignmentPairs}, or
\link[GenomicAlignments:GAlignmentsList]{GenomicAlignments::GAlignmentsList} object representing a set
of reads or read pairs.}

\item{fraglen}{The estimated fragment length. This is only required
if \code{reads} contains any un-paired reads, since the far end of
the fragment represented by a single-end read is unknown.}

\item{extend_fragment_upstream}{If FALSE (the default), each
single-end read is assumed to represent the 5-prime end of a
fragment, so the fragment will be extended \code{fraglen} base pairs
in the 3-prime direction starting from the 5-prime end of the
read. If reads represent the 3-prime ends of fragments, set
this argument to TRUE, and the reverse will happen: fragments
will be extended \code{fraglen} base pairs in the 5-prime direction
starting from the 3-prime end of the read.

This should really be an S4 method, but writing S4 methods is a
pain.}
}
\description{
Compute the midpoints of a collection of read fragments
}
