% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{save_image_filtered}
\alias{save_image_filtered}
\title{Variant of \code{save.image()} that allows excluding specific names}
\usage{
save_image_filtered(
  file = ".RData",
  version = NULL,
  ascii = FALSE,
  compress = !ascii,
  safe = TRUE,
  exclude = NULL
)
}
\arguments{
\item{file, version, ascii, compress, safe}{These arguments have the
same meaning as in \code{save.image()}.}

\item{exclude}{Character vector of variable names not to save.}
}
\description{
If you want to save your entire workspace except for a specific set
of variable names that you wish to exclude, this function makes it
easy to do so.
}
\examples{

x <- 5
y <- 3
datafile <- tempfile(fileext = ".rda")
# Only saves y, not x
save_image_filtered(datafile, exclude = "x")

}
\seealso{
\code{\link[=save.image]{save.image()}}
}
