% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{setup_multicore}
\alias{setup_multicore}
\title{Set up all parallel backends to use multicore}
\usage{
setup_multicore(ncores = future::availableCores())
}
\arguments{
\item{ncores}{The number of cores to use. If not specified, it is
set automatically from the environment, using
\code{\link[future:availableCores]{future::availableCores()}}.}
}
\description{
There are several different parallel-processing packages with
multiple backends. This function sets all of them up to use
multi-core processing.
}
