% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si2f.R
\name{si2f}
\alias{si2f}
\title{Inverse of \code{\link[sitools:f2si]{sitools::f2si()}}}
\usage{
si2f(string, unit = "")
}
\arguments{
\item{string}{A character vector representing numbers along with SI
prefixes and possibly a unit.}

\item{unit}{The unit to expect after the SI prefix, if any.}
}
\value{
A numeric vector containing the values parsed from
\code{string}.
}
\description{
Inverse of \code{\link[sitools:f2si]{sitools::f2si()}}
}
\examples{
# convert single number
si2f("10k")

# convert single number with unit
si2f("23 mV", unit="V")

# convert list of numbers
si_strings <- c("100 k", "35 E", "4 m")
si2f(si_strings)

}
