% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{use_futures}
\alias{use_futures}
\title{Set up foreach and BiocParallel to use futures}
\usage{
use_futures(strategy, ..., quiet = FALSE)
}
\arguments{
\item{strategy, ...}{If provided, these are passed directly to
\code{\link[future:plan]{future::plan()}}. If \code{strategy} is not provided,
\code{future::plan()} is not called at all, leaving the current
execution strategy unchanged and ignoring any additional
arguments. This is useful if you have already called
\code{future::plan()} and you just want to set up other parallel
packages to use futures.}

\item{quiet}{If FALSE (the default), indicate what is being done.
If TRUE, do not issue any messages.

Note that this forces loading of the BiocParallel and foreach
packages. Ideally, this function would only set up a hook to run
the appropriate setup code after these packages are loaded, thus
saving the setup time if they are never loaded. However, this does
not appear to be possible. The available "on load" hook mechanism
seems to only trigger when a package is attached, not when it is
loaded. Hence, if another pacakge uses BiocParallel or foreach
internally, (e.g. \code{\link[SummarizedExperiment:summarizeOverlaps]{SummarizedExperiment::summarizeOverlaps()}}),
this would not trigger the hook.}
}
\examples{

# Set up BiocParallel and foreach to use futures, keeping the
# existing future execution strategy.
use_futures()

# Same, but switch to the multicore strategy.
use_futures("multisession")

# Same, but pass additional options to the multicore strategy.
use_futures("multisession", workers = 2)

}
