% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{windowCountsParallel}
\alias{windowCountsParallel}
\title{Parallel version of \code{\link[csaw:windowCounts]{csaw::windowCounts()}}}
\usage{
windowCountsParallel(bam.files, ..., filter = 10,
  BPPARAM = BiocParallel::bpparam())
}
\description{
Parallel version of \code{\link[csaw:windowCounts]{csaw::windowCounts()}}
}
