% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_io.R
\name{write_narrowPeak}
\alias{write_narrowPeak}
\title{Write a narrowPeak format BED file.}
\usage{
write_narrowPeak(x, file, ...)
}
\arguments{
\item{x}{A GRanges or data frame containing the all appropriate
columns for a narrowPeak file. In addition to the normal BED
columns, it must contain columns named "signalValue", "pValue",
"qValue", and "summit".}

\item{file}{The file name to write to.}

\item{...}{Additional arguments are passed to \code{write.table()}.

If some of the required columns are not available (e.g. from a peak
caller that does not report a summit), you can fill in the
unavailable columns with NA.}
}
\description{
This function takes a GRanges of the kind returned by
\code{\link[=read_narrowPeak]{read_narrowPeak()}} and writes it back to a file.
}
