% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{cleanup_mcols}
\alias{cleanup_mcols}
\title{Remove valueless mcols from an object}
\usage{
cleanup_mcols(object, mcols_df)
}
\description{
Emptiness is defined as in \code{\link[=is_valueless]{is_valueless()}}.
}
