% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{code_control_named}
\alias{code_control_named}
\title{Variant of code_control that generates more verbose column names}
\usage{
code_control_named(n, contrasts = TRUE, sparse = FALSE)
}
\description{
Variant of code_control that generates more verbose column names
}
