% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{combineFCResults}
\alias{combineFCResults}
\title{Combine multiple featureCounts results}
\usage{
combineFCResults(fcreslist)
}
\description{
This is a helper function for \code{featureCountsParallel()}.
}
