% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes_for_builtins.R
\name{deparse_onestring}
\alias{deparse_onestring}
\title{Deparse and then concatenate into a single string}
\usage{
deparse_onestring(...)
}
\description{
Deparse and then concatenate into a single string
}
