% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_utils.R
\name{eBayes_auto_proportion}
\alias{eBayes_auto_proportion}
\title{Variant of eBayes that sets the proportion argument automatically.}
\usage{
eBayes_auto_proportion(..., prop.method = "lfdr")
}
\description{
Variant of eBayes that sets the proportion argument automatically.
}
