% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_utils.R
\name{estimateDispByGroup}
\alias{estimateDispByGroup}
\title{Estimate edgeR dispersions separately for each group}
\usage{
estimateDispByGroup(dge, group = as.factor(dge$samples$group), batch, ...)
}
\description{
Estimate edgeR dispersions separately for each group
}
