% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{format_bp}
\alias{format_bp}
\title{Format a number of base pairs using the most appropriate unit.}
\usage{
format_bp(x)
}
\description{
Format a number of base pairs using the most appropriate unit.
}
