% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{future.lapply}
\alias{future.lapply}
\title{Like lapply, but returns future objects.}
\usage{
future.lapply(X, FUN, ...)
}
\description{
The results can be fetched all at once with \code{\link[future:values]{future::values()}}.
}
