% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_utils.R
\name{get_pval_colname}
\alias{get_pval_colname}
\title{Determine the column name of the p-value column in a table}
\usage{
get_pval_colname(ttab)
}
\description{
limma uses "P.Value", edgeR uses "PValue", so we need an
abstraction.
}
