% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plots.R
\name{ggplotBCV}
\alias{ggplotBCV}
\title{ggplot version of \code{edgeR::plotBCV()}.}
\usage{
ggplotBCV(y, xlab = "Average log CPM",
  ylab = "Biological coefficient of variation", rawdisp = NULL, ...)
}
\description{
Additional arguments passed to \code{\link[=getBCVTable]{getBCVTable()}}.
}
