% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{load_in_new_env}
\alias{load_in_new_env}
\title{Load an R data file into a new environment}
\usage{
load_in_new_env(file, envir = new.env(), ...)
}
\arguments{
\item{file}{This argument has the same meaning as in \code{load()}}

\item{envir}{If you pass an existing environment, objects will be
loaded into that environment instead of a newly-created one,
and this same environment will be returned.}

\item{...}{Further arguments are passed to \code{load()}.}
}
\value{
An environment containing all the objects loaded from
\code{file}.
}
\description{
Unlike \code{load()}, this function avoids polluting the global
namespace by loading objects into separate environment.
}
\details{
This function is useful for loading data files that potentially
contain many objects with common names that might otherwise
overwrite existing objects with the same name.
}
\examples{

x <- 5
y <- 3
datafile <- tempfile(fileext = ".rda")
save(file = datafile, list = c("x", "y"))
rm(x,y)
loaded_vars <- load_in_new_env(datafile)
as.list(loaded_vars)

}
\seealso{
\code{\link[=save.image]{save.image()}}
}
