% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_utils.R
\name{make_futures}
\alias{make_futures}
\title{Construct futures for several expressions}
\usage{
make_futures(..., .future.args = list())
}
\arguments{
\item{...}{Expressions for which to construct futures.}

\item{.future.args}{Additional arguments to pass to
\code{\link[future:future]{future::future()}}.}
}
\value{
A list of futures, one for each expression in \code{...}.
}
\description{
Instead of evaluating its arguments, this function converts them
into futures using \code{\link[future:future]{future::future()}}. Calling \code{\link[future:value]{future::value()}} on
each one will yield the value of that expression.
}
\examples{

library(future)
flist <- make_futures(a = 1+1, b = 2+2)
flist
values(flist)

}
