% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes_for_builtins.R
\name{match_arg}
\alias{match_arg}
\title{Improved match.arg with automatic detection of the argument name}
\usage{
match_arg(arg, choices, several.ok = FALSE, arg.name = substitute(arg),
  ignore.case = FALSE)
}
\description{
Improved match.arg with automatic detection of the argument name
}
