% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_utils.R
\name{parse_bp}
\alias{parse_bp}
\title{Parse a number of base pairs with optional units.}
\usage{
parse_bp(size)
}
\description{
Parse a number of base pairs with optional units.
}
\examples{
parse_bp(c("100", "100bp", "100 kbp", "1Mbp"))

}
