% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_utils.R
\name{print_var_vector}
\alias{print_var_vector}
\title{Print a readable summary of a list of values.}
\usage{
print_var_vector(v)
}
\arguments{
\item{v}{A named list or vector}
}
\value{
Returns \code{v} invisibly.
}
\description{
This is useful for printing out a list of the parsed command-line
arguments for a script. The output is generally more compact than
\code{\link[=print]{print()}} or \code{\link[=str]{str()}}.
}
\examples{

x <- list(verbose = TRUE, infile = "a.txt", outfile = "b.txt", ncores = 8)
print_var_vector(x)

}
