% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{promote_common_mcols}
\alias{promote_common_mcols}
\title{Promote common mcols from the GRanges inside a GRangesList}
\usage{
promote_common_mcols(grl, delete.from.source = FALSE, ...)
}
\description{
Given a GRangesList whose underlying ranges have mcols, this finds
mcols of the ranges that are constant within each gene and promotes
them to mcols of the GRangesList. For example, you have a
GRangesList of exons grouped by gene, and the exons are annotated
with gene IDs, then the gene ID column will be promoted to the
GRangesList object itself.
}
