% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rctutils-package}
\alias{rctutils}
\alias{rctutils-package}
\title{Common imports}
\description{
This is just here to tell roxygen2 about all base package imports,
which were recommended by R CMD check, as well as some common
imports that are used in many functions. Adding these to every
individual function that uses these common functions is too
tedious, so I've just added them all here.
}
\author{
\strong{Maintainer}: Ryan C. Thompson \email{rct@thompsonclan.org}

}
