% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_io.R
\name{save_RDS_or_RDA}
\alias{save_RDS_or_RDA}
\title{Save a single object to am RDS or RDA file}
\usage{
save_RDS_or_RDA(object, file, ascii = FALSE, version = NULL,
  compress = TRUE, savetype)
}
\description{
If the output file name ends in ".rda" or ".rdata", the file will
be saved using \code{\link[=save]{save()}}. Otherwise, it will be saved using
\code{\link[=saveRDS]{saveRDS()}}.
}
