% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_utils.R
\name{subtractCoefs}
\alias{subtractCoefs}
\title{Alternative interface to \code{limma::removeBatchEffect()}}
\usage{
subtractCoefs(x, design, coefsToSubtract, ...)
}
\arguments{
\item{x}{This has the same meaning as in
\code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.}

\item{design}{The full design matrix, including both covariates
(whose effects are to be kept) and batch effects (whose effects
are to be subtracted out).}

\item{coefsToSubtract}{An index vector indicating which columns of
\code{design} should be subtracted out. This can be any vector such
that \code{design[,coefsToSubtract]} selects the appropriate
columns.}

\item{...}{Additional arguments are passed to
\code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.}
}
\value{
See \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.
}
\description{
Instead of accepting the batch effects and covariates separately,
this function accepts a single design matrix, and an index of which
columns of that design matrix represent batch effects.
}
\examples{

#TODO: Copy from removeBatchEffect

}
