% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_utils.R
\name{tsmsg}
\alias{tsmsg}
\title{Like \code{message} but with a timestamp}
\usage{
tsmsg(...)
}
\arguments{
\item{...}{All arguments are passed directly to \code{message}.

The time stamp is generated using \code{\link[=date]{date()}}. If you need anything
more complicated than this, it's probably easier to abondon this
function and just use \code{date()} and \code{message()} to manually add the
timestamp in the format you want.}
}
\description{
This is identical to the \code{\link[=message]{message()}} function, except that it
prepends a timestamp to every message.
}
\examples{

tsmsg("Hello world")

}
\seealso{
\code{\link[=message]{message()}}, \code{\link[=date]{date()}}
}
